Beginning upload note for this project
The following command was used in the creation of this boiler plate project of a login page:

dotnet new mvc --auth Individual -o loginMVC

Using the following websites:

https://docs.microsoft.com/en-us/aspnet/core/security/authentication/identity?view=aspnetcore-2.2&tabs=netcore-cli

Which I modify the command into a MVC project  from the guid provided in the following link:

https://docs.microsoft.com/en-us/dotnet/core/tools/dotnet-new?tabs=netcore21

Razor page projects I get a feeling seems to have a backwards feeling to them. It feels like the old LAMP project days, with HTML and PHP. It does follow the MVVM model, but it seems to mush the model and controller into one class file while the view is handled in the cshtml file.

Razor pages are relatively new, we'll see how popular it becomes if it warrants further study. For now I'll be following the MVC login page to get an idea how Identity works.